Move bank dependent tests out of serde_snapshot/tests.rs
