+++ runtime/src/bank.rs
@@ -209,6 +209,8 @@ struct VerifyAccountsHashConfig {
 mod address_lookup_table;
 mod builtin_programs;
 mod metrics;
+#[cfg(test)]
+mod serde_snapshot;
 mod sysvar_cache;
 #[cfg(test)]
 mod tests;
